test(testing): разделены крупные test suites
articles screen, settings, source management и feed refresh tests разнесены по behavior-focused файлам
общая test support логика для settings controller вынесена отдельно
