+++ RSSReaderTests/ArticlesScreen/ArticlesScreenControllerMarkAllReadTests.swift
@@ -1,0 +1,82 @@
+import Foundation
+import SwiftData
+import Testing
+@testable import RSSReader
+
+@Suite("Articles Screen / Controller / Mark All Read")
+@MainActor
+struct ArticlesScreenControllerMarkAllReadTests {
+    @Test
+    func articlesScreenControllerPresentsConfirmationWhenAskBeforeMarkingAllAsReadIsEnabled() throws {
+        let harness = try TestHarness.make(httpClient: ScriptedHTTPClient())
+        let appSettingsRepository = try #require(harness.dependencies.appSettingsRepository)
+        let unreadItem = makeArticleListItemDTO(isRead: false, isStarred: false)
+        let controller = ArticlesScreenController(
+            previewScreenState: .previewLoaded(
+                selection: .feed(unreadItem.feedID),
+                navigationTitle: "Feed",
+                navigationSubtitle: "1 Unread Item",
+                articles: [unreadItem]
+            )
+        )
+
+        _ = try appSettingsRepository.update(
+            AppSettingsUpdate(
+                askBeforeMarkingAllAsRead: true,
+                updatedAt: .distantPast
+            )
+        )
+
+        controller.handleMarkAllAsReadAction(
+            searchText: "",
+            selection: .feed(unreadItem.feedID),
+            sourcesFilter: .allItems,
+            dependencies: harness.dependencies,
+            isPreviewMode: false
+        )
+
+        #expect(controller.screenState.pendingConfirmation == .markAllAsRead)
+        #expect(controller.screenState.articles.first?.isRead == false)
+    }
+
+    @Test
+    func articlesScreenControllerMarksAllAsReadImmediatelyWhenAskBeforeMarkingAllAsReadIsDisabled() throws {
+        let harness = try TestHarness.make(httpClient: ScriptedHTTPClient())
+        let appSettingsRepository = try #require(harness.dependencies.appSettingsRepository)
+        let articleStateRepository = try #require(harness.dependencies.articleStateRepository)
+        let unreadItem = makeArticleListItemDTO(isRead: false, isStarred: false)
+        let controller = ArticlesScreenController(
+            previewScreenState: .previewLoaded(
+                selection: .feed(unreadItem.feedID),
+                navigationTitle: "Feed",
+                navigationSubtitle: "1 Unread Item",
+                articles: [unreadItem]
+            )
+        )
+
+        _ = try appSettingsRepository.update(
+            AppSettingsUpdate(
+                askBeforeMarkingAllAsRead: false,
+                updatedAt: .distantPast
+            )
+        )
+
+        controller.handleMarkAllAsReadAction(
+            searchText: "",
+            selection: .feed(unreadItem.feedID),
+            sourcesFilter: .allItems,
+            dependencies: harness.dependencies,
+            isPreviewMode: false
+        )
+
+        let persistedState = try articleStateRepository.fetchStateSnapshot(
+            feedID: unreadItem.feedID,
+            articleExternalID: unreadItem.articleExternalID
+        )
+
+        #expect(controller.screenState.pendingConfirmation == nil)
+        #expect(controller.screenState.articles.first?.isRead == true)
+        #expect(controller.screenState.navigationSubtitle == "No Unread Items")
+        #expect(persistedState?.isRead == true)
+    }
+}
