+++ RSSReaderTests/Feeds/FeedRefreshServiceSingleFeedTests.swift
@@ -1,0 +1,241 @@
+import Foundation
+import Testing
+@testable import RSSReader
+
+@Suite("Feeds / Refresh Service / Single Feed")
+@MainActor
+struct FeedRefreshServiceSingleFeedTests {
+    @Test
+    func singleFeedRefreshFetchedPersistsArticlesMetadataAndFetchState() async throws {
+        let oldSuccessAt = Date(timeIntervalSince1970: 1_700_000_000)
+        let harness = try TestHarness.make(
+            httpClient: ScriptedHTTPClient(
+                steps: [
+                    .response(
+                        statusCode: 200,
+                        headers: [
+                            "Content-Type": "application/rss+xml; charset=utf-8",
+                            "ETag": "\"etag-new\"",
+                            "Last-Modified": "Tue, 02 Jan 2024 12:00:00 GMT"
+                        ],
+                        body: makeValidRSSFeedXML(
+                            channelTitle: "Updated Feed Title",
+                            channelLink: "https://example.com/",
+                            language: "en",
+                            itemTitle: "Article One",
+                            itemLink: "https://example.com/articles/1",
+                            itemGUID: "article-1",
+                            itemDescription: "Readable summary",
+                            pubDate: "Tue, 02 Jan 2024 10:00:00 GMT"
+                        )
+                    )
+                ]
+            )
+        )
+
+        let feed = Feed(
+            url: "https://example.com/feed.xml",
+            title: "Old Feed Title",
+            lastSuccessfulFetchAt: oldSuccessAt,
+            lastETag: "\"etag-old\"",
+            lastModifiedHeader: "Mon, 01 Jan 2024 12:00:00 GMT",
+            lastSyncError: "Previous error"
+        )
+        try harness.feedRepository.insert(feed)
+
+        let result = await harness.service.refresh(feedID: feed.id)
+
+        #expect(result.status == .fetched)
+        #expect(result.processedEntryCount == 1)
+        #expect(result.upsertedEntryCount == 1)
+        #expect(result.rejectedEntryCount == 0)
+        #expect(result.errorDescription == nil)
+
+        let fetchedFeed = try harness.fetchFeed(id: feed.id)
+        let refreshedFeed = try #require(fetchedFeed)
+        #expect(refreshedFeed.title == "Updated Feed Title")
+        #expect(refreshedFeed.siteURL == "https://example.com/")
+        #expect(refreshedFeed.language == "en")
+        #expect(refreshedFeed.kind == .rss)
+        #expect(refreshedFeed.lastFetchedAt != nil)
+        #expect(refreshedFeed.lastSuccessfulFetchAt != nil)
+        #expect(refreshedFeed.lastSuccessfulFetchAt != oldSuccessAt)
+        #expect(refreshedFeed.lastETag == "\"etag-new\"")
+        #expect(refreshedFeed.lastModifiedHeader == "Tue, 02 Jan 2024 12:00:00 GMT")
+        #expect(refreshedFeed.lastSyncError == nil)
+
+        let articles = try harness.articleRepository.fetchArticles(feedID: feed.id)
+        #expect(articles.count == 1)
+        #expect(articles.first?.title == "Article One")
+
+        let fetchedLog = try harness.feedFetchLogRepository.fetchLatestLog(feedID: feed.id)
+        let latestLog = try #require(fetchedLog)
+        #expect(latestLog.status == "fetched")
+        #expect(latestLog.httpCode == 200)
+
+        let requests = await harness.httpClient.recordedRequests()
+        #expect(requests.count == 1)
+        #expect(requests.first?.headers["If-None-Match"] == "\"etag-old\"")
+        #expect(requests.first?.headers["If-Modified-Since"] == "Mon, 01 Jan 2024 12:00:00 GMT")
+    }
+
+    @Test
+    func singleFeedRefreshNotModifiedUpdatesFetchStateWithoutParsingPipeline() async throws {
+        let oldSuccessAt = Date(timeIntervalSince1970: 1_700_000_000)
+        let harness = try TestHarness.make(
+            httpClient: ScriptedHTTPClient(
+                steps: [
+                    .response(
+                        statusCode: 304,
+                        headers: [
+                            "ETag": "\"etag-304\"",
+                            "Last-Modified": "Wed, 03 Jan 2024 12:00:00 GMT"
+                        ],
+                        body: ""
+                    )
+                ]
+            )
+        )
+
+        let feed = Feed(
+            url: "https://example.com/feed.xml",
+            title: "Stable Feed Title",
+            lastSuccessfulFetchAt: oldSuccessAt,
+            lastETag: "\"etag-old\"",
+            lastModifiedHeader: "Mon, 01 Jan 2024 12:00:00 GMT",
+            lastSyncError: "Transient error"
+        )
+        try harness.feedRepository.insert(feed)
+
+        let result = await harness.service.refresh(feedID: feed.id)
+
+        #expect(result.status == .notModified)
+        #expect(result.processedEntryCount == 0)
+        #expect(result.upsertedEntryCount == 0)
+        #expect(result.rejectedEntryCount == 0)
+        #expect(result.errorDescription == nil)
+
+        let fetchedFeed = try harness.fetchFeed(id: feed.id)
+        let refreshedFeed = try #require(fetchedFeed)
+        #expect(refreshedFeed.title == "Stable Feed Title")
+        #expect(refreshedFeed.lastFetchedAt != nil)
+        #expect(refreshedFeed.lastSuccessfulFetchAt == oldSuccessAt)
+        #expect(refreshedFeed.lastETag == "\"etag-304\"")
+        #expect(refreshedFeed.lastModifiedHeader == "Wed, 03 Jan 2024 12:00:00 GMT")
+        #expect(refreshedFeed.lastSyncError == nil)
+
+        let articles = try harness.articleRepository.fetchArticles(feedID: feed.id)
+        #expect(articles.isEmpty)
+
+        let fetchedLog = try harness.feedFetchLogRepository.fetchLatestLog(feedID: feed.id)
+        let latestLog = try #require(fetchedLog)
+        #expect(latestLog.status == "not_modified")
+        #expect(latestLog.httpCode == 304)
+
+        let requests = await harness.httpClient.recordedRequests()
+        #expect(requests.count == 1)
+        #expect(requests.first?.headers["If-None-Match"] == "\"etag-old\"")
+        #expect(requests.first?.headers["If-Modified-Since"] == "Mon, 01 Jan 2024 12:00:00 GMT")
+    }
+
+    @Test
+    func singleFeedRefreshFailedPersistsErrorWithoutWritingArticles() async throws {
+        let harness = try TestHarness.make(
+            httpClient: ScriptedHTTPClient(
+                steps: [
+                    .response(
+                        statusCode: 500,
+                        headers: [
+                            "Content-Type": "application/rss+xml; charset=utf-8"
+                        ],
+                        body: ""
+                    )
+                ]
+            )
+        )
+
+        let oldSuccessAt = Date(timeIntervalSince1970: 1_700_000_000)
+        let feed = Feed(
+            url: "https://example.com/feed.xml",
+            title: "Failing Feed",
+            lastSuccessfulFetchAt: oldSuccessAt,
+            lastETag: "\"etag-old\"",
+            lastModifiedHeader: "Mon, 01 Jan 2024 12:00:00 GMT"
+        )
+        try harness.feedRepository.insert(feed)
+
+        let result = await harness.service.refresh(feedID: feed.id)
+
+        #expect(result.status == .failed)
+        #expect(result.processedEntryCount == 0)
+        #expect(result.upsertedEntryCount == 0)
+        #expect(result.rejectedEntryCount == 0)
+        #expect(result.errorDescription?.contains("invalidStatusCode") == true)
+
+        let fetchedFeed = try harness.fetchFeed(id: feed.id)
+        let refreshedFeed = try #require(fetchedFeed)
+        #expect(refreshedFeed.title == "Failing Feed")
+        #expect(refreshedFeed.lastFetchedAt != nil)
+        #expect(refreshedFeed.lastSuccessfulFetchAt == oldSuccessAt)
+        #expect(refreshedFeed.lastSyncError?.contains("invalidStatusCode") == true)
+        #expect(refreshedFeed.lastETag == "\"etag-old\"")
+        #expect(refreshedFeed.lastModifiedHeader == "Mon, 01 Jan 2024 12:00:00 GMT")
+
+        let articles = try harness.articleRepository.fetchArticles(feedID: feed.id)
+        #expect(articles.isEmpty)
+
+        let fetchedLog = try harness.feedFetchLogRepository.fetchLatestLog(feedID: feed.id)
+        let latestLog = try #require(fetchedLog)
+        #expect(latestLog.status == "failed")
+        #expect(latestLog.httpCode == 500)
+        #expect(latestLog.message?.contains("invalidStatusCode") == true)
+    }
+
+    @Test
+    func singleFeedRefreshCancelledReturnsCancelledWithoutPersistingFailureState() async throws {
+        let oldSuccessAt = Date(timeIntervalSince1970: 1_700_000_000)
+        let harness = try TestHarness.make(
+            httpClient: ScriptedHTTPClient(
+                steps: [
+                    .cancelled
+                ]
+            )
+        )
+
+        let feed = Feed(
+            url: "https://example.com/feed.xml",
+            title: "Cancellable Feed",
+            lastSuccessfulFetchAt: oldSuccessAt,
+            lastETag: "\"etag-old\"",
+            lastModifiedHeader: "Mon, 01 Jan 2024 12:00:00 GMT",
+            lastSyncError: "Previous error"
+        )
+        try harness.feedRepository.insert(feed)
+
+        let result = await harness.service.refresh(feedID: feed.id)
+
+        #expect(result.status == .cancelled)
+        #expect(result.processedEntryCount == 0)
+        #expect(result.upsertedEntryCount == 0)
+        #expect(result.rejectedEntryCount == 0)
+        #expect(result.errorDescription == "Refresh cancelled")
+
+        let fetchedFeed = try harness.fetchFeed(id: feed.id)
+        let refreshedFeed = try #require(fetchedFeed)
+        #expect(refreshedFeed.title == "Cancellable Feed")
+        #expect(refreshedFeed.lastFetchedAt != nil)
+        #expect(refreshedFeed.lastSuccessfulFetchAt == oldSuccessAt)
+        #expect(refreshedFeed.lastSyncError == "Previous error")
+        #expect(refreshedFeed.lastETag == "\"etag-old\"")
+        #expect(refreshedFeed.lastModifiedHeader == "Mon, 01 Jan 2024 12:00:00 GMT")
+
+        let articles = try harness.articleRepository.fetchArticles(feedID: feed.id)
+        #expect(articles.isEmpty)
+
+        let fetchedLog = try harness.feedFetchLogRepository.fetchLatestLog(feedID: feed.id)
+        let latestLog = try #require(fetchedLog)
+        #expect(latestLog.status == "cancelled")
+        #expect(latestLog.httpCode == nil)
+        #expect(latestLog.message?.contains("Refresh cancelled") == true)
+    }
+}
