+++ RSSReaderTests/SourceManagement/SourceManagementScreenControllerAddFeedTests.swift
@@ -2,9 +2,9 @@ import Foundation
 import Testing
 @testable import RSSReader
 
-@Suite("Source Management / Screen Controller")
+@Suite("Source Management / Screen Controller / Add Feed")
 @MainActor
-struct SourceManagementScreenControllerTests {
+struct SourceManagementScreenControllerAddFeedTests {
     @Test
     func sourceManagementScreenControllerLoadsFeedPreviewThroughSourceManagementService() async throws {
         let feedURL = "https://example.com/preview-controller.xml"
@@ -263,80 +263,6 @@ struct SourceManagementScreenControllerTests {
     }
 
     @Test
-    func sourceManagementScreenControllerEditsFolderFromSidebarLaunchContextAndRetargetsSelection() throws {
-        let harness = try TestHarness.make(httpClient: ScriptedHTTPClient())
-        let appState = AppState()
-        let folder = try harness.folderRepository.insert(Folder(name: "News", sortOrder: 0))
-        let controller = SourceManagementScreenController()
-        let sidebarReloadIDBeforeEdit = appState.sourcesSidebarReloadID
-
-        harness.dependencies.appActions.showFolder(named: "News", using: appState)
-        harness.dependencies.appActions.showFolderEditor(named: "News", using: appState)
-        controller.handleLaunchContext(.editFolder(folder.id), dependencies: harness.dependencies)
-
-        guard case .createFolder(let initialDestination)? = controller.viewState().presentedDestination else {
-            Issue.record("Expected create-folder destination presentation for folder editor launch context")
-            return
-        }
-
-        #expect(initialDestination.title == "Edit Folder")
-        #expect(initialDestination.nameInput == "News")
-
-        controller.handleCreateFolderNameChange("World News")
-        controller.submitCreateFolder(dependencies: harness.dependencies, appState: appState)
-
-        let renamedFolder = try #require(try harness.folderRepository.fetchFolder(id: folder.id))
-
-        #expect(appState.isPresentingSourceManagementScreen == false)
-        #expect(appState.selectedSidebarSelection == .folder("World News"))
-        #expect(appState.sourcesSidebarReloadID != sidebarReloadIDBeforeEdit)
-        #expect(renamedFolder.name == "World News")
-        #expect(renamedFolder.sortOrder == 0)
-    }
-
-    @Test
-    func sourceManagementScreenControllerOrganizesFeedFromSidebarLaunchContextWithoutPreview() async throws {
-        let harness = try TestHarness.make(httpClient: ScriptedHTTPClient())
-        let appState = AppState()
-        let newsFolder = try harness.folderRepository.insert(Folder(name: "News", sortOrder: 0))
-        let techFolder = try harness.folderRepository.insert(Folder(name: "Tech", sortOrder: 1))
-        let feed = try harness.feedRepository.insert(
-            Feed(
-                url: "https://example.com/organize-me.xml",
-                title: "Organize Me",
-                kind: .rss,
-                folder: newsFolder
-            )
-        )
-        let controller = SourceManagementScreenController()
-        let sidebarReloadIDBeforeMove = appState.sourcesSidebarReloadID
-
-        harness.dependencies.appActions.showFeedOrganizer(id: feed.id, using: appState)
-        controller.handleLaunchContext(.organizeFeed(feed.id), dependencies: harness.dependencies)
-
-        guard case .moveSource(let initialDestination)? = controller.viewState().presentedDestination else {
-            Issue.record("Expected move-source destination presentation for feed organizer launch context")
-            return
-        }
-
-        #expect(appState.sourceManagementLaunchContext == .organizeFeed(feed.id))
-        #expect(initialDestination.feeds.first(where: { $0.id == feed.id })?.isSelected == true)
-        #expect(initialDestination.placementOptions.first(where: { $0.title == "News" })?.isSelected == true)
-        #expect(initialDestination.isPrimaryActionEnabled == false)
-
-        controller.handleMoveSourcePlacementSelection(.folder(techFolder.id))
-        controller.submitMoveSource(dependencies: harness.dependencies, appState: appState)
-
-        let persistedFeed = try #require(try harness.feedRepository.fetchFeed(id: feed.id))
-        let requests = await harness.httpClient.recordedRequests()
-
-        #expect(appState.isPresentingSourceManagementScreen == false)
-        #expect(appState.sourcesSidebarReloadID != sidebarReloadIDBeforeMove)
-        #expect(persistedFeed.folder?.id == techFolder.id)
-        #expect(requests.isEmpty)
-    }
-
-    @Test
     func sourceManagementScreenControllerShowsDuplicateFeedWarningWhenPreviewMatchesExistingSource() async throws {
         let feedURL = "https://example.com/existing-feed.xml"
         let harness = try TestHarness.make(
@@ -445,264 +371,5 @@ struct SourceManagementScreenControllerTests {
         #expect(destination.status?.detail == "The address responded with text/html; charset=utf-8, not a supported RSS or Atom feed.")
         #expect(destination.primaryActionTitle == "Preview Feed")
         #expect(destination.isPrimaryActionEnabled == false)
-    }
-
-    @Test
-    func sourceManagementScreenControllerMovesFeedThroughServiceAndRefreshesDestination() throws {
-        let harness = try TestHarness.make(httpClient: ScriptedHTTPClient())
-        let newsFolder = try harness.folderRepository.insert(Folder(name: "News", sortOrder: 0))
-        let techFolder = try harness.folderRepository.insert(Folder(name: "Tech", sortOrder: 1))
-        let feed = try harness.feedRepository.insert(
-            Feed(
-                url: "https://example.com/move-me.xml",
-                title: "Move Me",
-                kind: .rss,
-                folder: newsFolder
-            )
-        )
-        let controller = SourceManagementScreenController()
-
-        controller.handleScenarioSelection(.moveSource, dependencies: harness.dependencies)
-
-        guard case .moveSource(let initialDestination)? = controller.viewState().presentedDestination else {
-            Issue.record("Expected move-source destination presentation")
-            return
-        }
-
-        #expect(initialDestination.feeds.map(\.title) == ["Move Me"])
-        #expect(initialDestination.feeds.first?.currentPlacementTitle == "News")
-        #expect(initialDestination.isPrimaryActionEnabled == false)
-
-        controller.handleMoveSourcePlacementSelection(.folder(techFolder.id))
-        controller.submitMoveSource(dependencies: harness.dependencies)
-
-        guard case .moveSource(let movedDestination)? = controller.viewState().presentedDestination else {
-            Issue.record("Expected move-source destination presentation after move")
-            return
-        }
-
-        let persistedFeed = try harness.feedRepository.fetchFeed(id: feed.id)
-
-        #expect(movedDestination.feedback?.kind == .success)
-        #expect(movedDestination.feedback?.detail?.contains("Tech") == true)
-        #expect(movedDestination.feeds.first?.currentPlacementTitle == "Tech")
-        #expect(movedDestination.isPrimaryActionEnabled == false)
-        #expect(persistedFeed?.folder?.id == techFolder.id)
-    }
-
-    @Test
-    func sourceManagementScreenControllerMovesFeedDismissesModalAndReloadsAffectedSelectionInAppFlow() throws {
-        let harness = try TestHarness.make(httpClient: ScriptedHTTPClient())
-        let newsFolder = try harness.folderRepository.insert(Folder(name: "News", sortOrder: 0))
-        let techFolder = try harness.folderRepository.insert(Folder(name: "Tech", sortOrder: 1))
-        let feed = try harness.feedRepository.insert(
-            Feed(
-                url: "https://example.com/app-move.xml",
-                title: "App Move",
-                kind: .rss,
-                folder: newsFolder
-            )
-        )
-        let controller = SourceManagementScreenController()
-        let appState = AppState()
-
-        harness.dependencies.appActions.showFolder(named: "News", using: appState)
-        let articleReloadIDBeforeMove = appState.articleListReloadID
-        let sidebarReloadIDBeforeMove = appState.sourcesSidebarReloadID
-
-        harness.dependencies.appActions.showSourceManagement(using: appState)
-        controller.handleScenarioSelection(.moveSource, dependencies: harness.dependencies)
-        controller.handleMoveSourcePlacementSelection(.folder(techFolder.id))
-        controller.submitMoveSource(dependencies: harness.dependencies, appState: appState)
-
-        let persistedFeed = try harness.feedRepository.fetchFeed(id: feed.id)
-
-        #expect(appState.isPresentingSourceManagementScreen == false)
-        #expect(appState.sourcesSidebarReloadID != sidebarReloadIDBeforeMove)
-        #expect(appState.articleListReloadID != articleReloadIDBeforeMove)
-        #expect(persistedFeed?.folder?.id == techFolder.id)
-    }
-
-    @Test
-    func sourceManagementScreenControllerReturnsToAddFeedWithNewFolderSelectedAfterInlineFolderCreation() async throws {
-        let feedURL = "https://example.com/feed.xml"
-        let harness = try TestHarness.make(
-            httpClient: ScriptedHTTPClient(
-                responsesByURL: [
-                    feedURL: .response(
-                        statusCode: 200,
-                        headers: ["Content-Type": "application/rss+xml; charset=utf-8"],
-                        body: makeValidRSSFeedXML(
-                            channelTitle: "Example Feed",
-                            channelLink: "https://example.com/",
-                            language: "en",
-                            itemTitle: "Example Article",
-                            itemLink: "https://example.com/articles/example",
-                            itemGUID: "example-article",
-                            itemDescription: "Example description",
-                            pubDate: "Tue, 02 Jan 2024 10:00:00 GMT"
-                        )
-                    )
-                ]
-            )
-        )
-        _ = try harness.folderRepository.insert(Folder(name: "News", sortOrder: 0))
-        let controller = SourceManagementScreenController()
-
-        controller.handleScenarioSelection(.addFeed, dependencies: harness.dependencies)
-        controller.handleAddFeedURLChange(feedURL)
-        await controller.handleAddFeedPrimaryAction(dependencies: harness.dependencies)
-
-        guard case .addFeed(let previewBeforeFolderDestination)? = controller.viewState().presentedDestination else {
-            Issue.record("Expected add-feed destination presentation after loading preview")
-            return
-        }
-
-        #expect(previewBeforeFolderDestination.urlInput == feedURL)
-        #expect(previewBeforeFolderDestination.primaryActionTitle == "Add Feed")
-        #expect(previewBeforeFolderDestination.isConfirmationActionEnabled)
-        #expect(previewBeforeFolderDestination.preview?.title == "Example Feed")
-
-        controller.startCreateFolderFromAddFeed(dependencies: harness.dependencies)
-
-        guard case .createFolder(let createFolderDestination)? = controller.viewState().presentedDestination else {
-            Issue.record("Expected create-folder destination when starting from add-feed flow")
-            return
-        }
-
-        #expect(createFolderDestination.existingFolders.map(\.name) == ["News"])
-
-        controller.handleCreateFolderNameChange("Research")
-        controller.submitCreateFolder(dependencies: harness.dependencies)
-
-        guard case .addFeed(let addFeedDestination)? = controller.viewState().presentedDestination else {
-            Issue.record("Expected add-feed destination after inline folder creation")
-            return
-        }
-
-        #expect(addFeedDestination.urlInput == feedURL)
-        #expect(addFeedDestination.primaryActionTitle == "Add Feed")
-        #expect(addFeedDestination.isConfirmationActionEnabled)
-        #expect(addFeedDestination.preview?.title == "Example Feed")
-        #expect(addFeedDestination.placementOptions.map(\.title) == ["Ungrouped", "News", "Research"])
-        #expect(addFeedDestination.placementOptions.last?.isSelected == true)
-    }
-
-    @Test
-    func sourceManagementScreenControllerRestoresPreviewedAddFeedAfterBackingOutOfNestedFolderCreation() async throws {
-        let feedURL = "https://example.com/feed.xml"
-        let harness = try TestHarness.make(
-            httpClient: ScriptedHTTPClient(
-                responsesByURL: [
-                    feedURL: .response(
-                        statusCode: 200,
-                        headers: ["Content-Type": "application/rss+xml; charset=utf-8"],
-                        body: makeValidRSSFeedXML(
-                            channelTitle: "Example Feed",
-                            channelLink: "https://example.com/",
-                            language: "en",
-                            itemTitle: "Example Article",
-                            itemLink: "https://example.com/articles/example",
-                            itemGUID: "example-article",
-                            itemDescription: "Example description",
-                            pubDate: "Tue, 02 Jan 2024 10:00:00 GMT"
-                        )
-                    )
-                ]
-            )
-        )
-        let newsFolder = try harness.folderRepository.insert(Folder(name: "News", sortOrder: 0))
-        let controller = SourceManagementScreenController()
-
-        controller.handleScenarioSelection(.addFeed, dependencies: harness.dependencies)
-        controller.handleAddFeedURLChange(feedURL)
-        await controller.handleAddFeedPrimaryAction(dependencies: harness.dependencies)
-        controller.handleAddFeedFolderPlacementSelection(.folder(newsFolder.id))
-        controller.startCreateFolderFromAddFeed(dependencies: harness.dependencies)
-        controller.dismissPresentedScenario()
-
-        guard case .addFeed(let addFeedDestination)? = controller.viewState().presentedDestination else {
-            Issue.record("Expected add-feed destination after backing out of nested folder creation")
-            return
-        }
-
-        #expect(addFeedDestination.urlInput == feedURL)
-        #expect(addFeedDestination.primaryActionTitle == "Add Feed")
-        #expect(addFeedDestination.isConfirmationActionEnabled)
-        #expect(addFeedDestination.preview?.title == "Example Feed")
-        #expect(addFeedDestination.placementOptions.first(where: { $0.title == "News" })?.isSelected == true)
-    }
-
-    @Test
-    func sourceManagementScreenControllerCreatesFolderThroughServiceAndRefreshesDestination() throws {
-        let harness = try TestHarness.make(httpClient: ScriptedHTTPClient())
-        let controller = SourceManagementScreenController()
-
-        _ = try harness.folderRepository.insert(Folder(name: "News", sortOrder: 0))
-
-        controller.handleScenarioSelection(.createFolder, dependencies: harness.dependencies)
-
-        guard case .createFolder(let initialDestination)? = controller.viewState().presentedDestination else {
-            Issue.record("Expected create-folder destination presentation")
-            return
-        }
-
-        #expect(initialDestination.existingFolders.map(\.name) == ["News"])
-        #expect(initialDestination.isPrimaryActionEnabled == false)
-
-        controller.handleCreateFolderNameChange("Research")
-
-        guard case .createFolder(let draftDestination)? = controller.viewState().presentedDestination else {
-            Issue.record("Expected create-folder destination presentation after input")
-            return
-        }
-
-        #expect(draftDestination.validationMessage == nil)
-        #expect(draftDestination.isPrimaryActionEnabled)
-
-        controller.submitCreateFolder(dependencies: harness.dependencies)
-
-        guard case .createFolder(let createdDestination)? = controller.viewState().presentedDestination else {
-            Issue.record("Expected create-folder destination presentation after submission")
-            return
-        }
-
-        #expect(createdDestination.nameInput.isEmpty)
-        #expect(createdDestination.existingFolders.map(\.name) == ["News", "Research"])
-        #expect(createdDestination.placementDescription == "This folder will be added after 2 existing folders.")
-        #expect(createdDestination.feedback?.kind == .success)
-
-        let folders = try harness.folderRepository.fetchAllFolders()
-        #expect(folders.map(\.name) == ["News", "Research"])
-        #expect(folders.map(\.sortOrder) == [0, 1])
-    }
-
-    @Test
-    func sourceManagementScreenControllerCreatesFolderRequestsSidebarReloadInAppFlow() throws {
-        let harness = try TestHarness.make(httpClient: ScriptedHTTPClient())
-        let controller = SourceManagementScreenController()
-        let appState = AppState()
-        let sidebarReloadIDBeforeCreation = appState.sourcesSidebarReloadID
-        let articleReloadIDBeforeCreation = appState.articleListReloadID
-
-        harness.dependencies.appActions.showSourceManagement(using: appState)
-        controller.handleScenarioSelection(.createFolder, dependencies: harness.dependencies)
-        controller.handleCreateFolderNameChange("Research")
-        controller.submitCreateFolder(
-            dependencies: harness.dependencies,
-            appState: appState
-        )
-
-        guard case .createFolder(let createdDestination)? = controller.viewState().presentedDestination else {
-            Issue.record("Expected create-folder destination presentation after app-level creation")
-            return
-        }
-
-        #expect(appState.isPresentingSourceManagementScreen)
-        #expect(appState.sourcesSidebarReloadID != sidebarReloadIDBeforeCreation)
-        #expect(appState.articleListReloadID == articleReloadIDBeforeCreation)
-        #expect(createdDestination.feedback?.title == "Folder created")
-        #expect(createdDestination.feedback?.detail == "\"Research\" is ready for sources.")
-        #expect(try harness.folderRepository.fetchFolder(name: "Research") != nil)
     }
 }
